+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - Add buttons for adding new server entries to the server list menu
 - Add ig.ENTITY.XenoDialog state
+- Add popup about problems when NW.js is outdated
 
 ### Fixed